backend/rector-api-main/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
pydantic>=2.4.2 
uagents>=0.22.3
google.generativeai>=0.6.0
firecrawl>=0.1.
langchain_google_genai>=0.0.5
langchain>=0.1.0